+++ Cider Remote/Components/ListHelper.swift
@@ -1,0 +1,35 @@
+// Made by Lumaa
+
+import SwiftUI
+
+struct GuideStep: View {
+    let number: Int
+    let text: String
+    
+    var body: some View {
+        HStack(alignment: .top, spacing: 15) {
+            Text("\(number)")
+                .font(.headline)
+                .foregroundColor(.white)
+                .frame(width: 30, height: 30)
+                .background(Circle().fill(Color.blue))
+            
+            Text(text)
+        }
+    }
+}
+
+struct BulletedList: View {
+    let items: [String]
+    
+    var body: some View {
+        VStack(alignment: .leading, spacing: 5) {
+            ForEach(items, id: \.self) { item in
+                HStack(alignment: .top, spacing: 10) {
+                    Text("•")
+                    Text(item)
+                }
+            }
+        }
+    }
+}
